xrscipy.signal.frequency_filter: support dask
The filter is applied along the core dimension. Supporting dask along
that dimension is non-trivial. It is however easy to support dask along
the non-core dimensions.

My use case is that I have lots of audio data that needs processing.
With this change, it is possible to apply the same filter to all channels,
and perform the computations out of core.